+++ lib/packages/wikibase/data-model/src/Term/AliasesProvider.php
@@ -3,7 +3,7 @@
 namespace Wikibase\DataModel\Term;
 
 /**
- * @since 4.0
+ * @since 4.1
  *
  * @licence GNU GPL v2+
  * @author Bene* < benestar.wikimedia@gmail.com >
@@ -13,6 +13,6 @@ interface AliasesProvider {
 	/**
 	 * @return AliasGroupList
 	 */
-	public function getAliases();
+	public function getAliasGroups();
 
 }